# proto/gen/api/v1/chat_service.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.11
// 	protoc        (unknown)
// source: api/v1/chat_service.proto

package apiv1

import (
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	_ "google.golang.org/protobuf/types/known/emptypb"
	_ "google.golang.org/protobuf/types/known/fieldmaskpb"
	_ "google.golang.org/protobuf/types/known/timestamppb"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type ChatRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Message       string                 `protobuf:"bytes,1,opt,name=message,proto3" json:"message,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ChatRequest) Reset() {
	*x = ChatRequest{}
	mi := &file_api_v1_chat_service_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ChatRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ChatRequest) ProtoMessage() {}

func (x *ChatRequest) ProtoReflect() protoreflect.Message {
	mi := &file_api_v1_chat_service_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ChatRequest.ProtoReflect.Descriptor instead.
func (*ChatRequest) Descriptor() ([]byte, []int) {
	return file_api_v1_chat_service_proto_rawDescGZIP(), []int{0}
}

func (x *ChatRequest) GetMessage() string {
	if x != nil {
		return x.Message
	}
	return ""
}

type ChatResponse struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Content       string                 `protobuf:"bytes,1,opt,name=content,proto3" json:"content,omitempty"`
	Type          string                 `protobuf:"bytes,2,opt,name=type,proto3" json:"type,omitempty"` // 思考内容 正式文本内容 搜索内容等等
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ChatResponse) Reset() {
	*x = ChatResponse{}
	mi := &file_api_v1_chat_service_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ChatResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ChatResponse) ProtoMessage() {}

func (x *ChatResponse) ProtoReflect() protoreflect.Message {
	mi := &file_api_v1_chat_service_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ChatResponse.ProtoReflect.Descriptor instead.
func (*ChatResponse) Descriptor() ([]byte, []int) {
	return file_api_v1_chat_service_proto_rawDescGZIP(), []int{1}
}

func (x *ChatResponse) GetContent() string {
	if x != nil {
		return x.Content
	}
	return ""
}

func (x *ChatResponse) GetType() string {
	if x != nil {
		return x.Type
	}
	return ""
}

var File_api_v1_chat_service_proto protoreflect.FileDescriptor

const file_api_v1_chat_service_proto_rawDesc = "" +
	"\n" +
	"\x19api/v1/chat_service.proto\x12\fmemos.api.v1\x1a\x1cgoogle/api/annotations.proto\x1a\x17google/api/client.proto\x1a\x1fgoogle/api/field_behavior.proto\x1a\x19google/api/resource.proto\x1a\x1bgoogle/protobuf/empty.proto\x1a google/protobuf/field_mask.proto\x1a\x1fgoogle/protobuf/timestamp.proto\"'\n" +
	"\vChatRequest\x12\x18\n" +
	"\amessage\x18\x01 \x01(\tR\amessage\"<\n" +
	"\fChatResponse\x12\x18\n" +
	"\acontent\x18\x01 \x01(\tR\acontent\x12\x12\n" +
	"\x04type\x18\x02 \x01(\tR\x04type2s\n" +
	"\vChatService\x12d\n" +
	"\x04Chat\x12\x19.memos.api.v1.ChatRequest\x1a\x1a.memos.api.v1.ChatResponse\"#\x82\xd3\xe4\x93\x02\x1d:\x01*\"\x18/api/v1/chat/completions0\x01B\xa8\x01\n" +
	"\x10com.memos.api.v1B\x10ChatServiceProtoP\x01Z0github.com/usememos/memos/proto/gen/api/v1;apiv1\xa2\x02\x03MAX\xaa\x02\fMemos.Api.V1\xca\x02\fMemos\\Api\\V1\xe2\x02\x18Memos\\Api\\V1\\GPBMetadata\xea\x02\x0eMemos::Api::V1b\x06proto3"

var (
	file_api_v1_chat_service_proto_rawDescOnce sync.Once
	file_api_v1_chat_service_proto_rawDescData []byte
)

func file_api_v1_chat_service_proto_rawDescGZIP() []byte {
	file_api_v1_chat_service_proto_rawDescOnce.Do(func() {
		file_api_v1_chat_service_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_api_v1_chat_service_proto_rawDesc), len(file_api_v1_chat_service_proto_rawDesc)))
	})
	return file_api_v1_chat_service_proto_rawDescData
}

var file_api_v1_chat_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_api_v1_chat_service_proto_goTypes = []any{
	(*ChatRequest)(nil),  // 0: memos.api.v1.ChatRequest
	(*ChatResponse)(nil), // 1: memos.api.v1.ChatResponse
}
var file_api_v1_chat_service_proto_depIdxs = []int32{
	0, // 0: memos.api.v1.ChatService.Chat:input_type -> memos.api.v1.ChatRequest
	1, // 1: memos.api.v1.ChatService.Chat:output_type -> memos.api.v1.ChatResponse
	1, // [1:2] is the sub-list for method output_type
	0, // [0:1] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_api_v1_chat_service_proto_init() }
func file_api_v1_chat_service_proto_init() {
	if File_api_v1_chat_service_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_api_v1_chat_service_proto_rawDesc), len(file_api_v1_chat_service_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_api_v1_chat_service_proto_goTypes,
		DependencyIndexes: file_api_v1_chat_service_proto_depIdxs,
		MessageInfos:      file_api_v1_chat_service_proto_msgTypes,
	}.Build()
	File_api_v1_chat_service_proto = out.File
	file_api_v1_chat_service_proto_goTypes = nil
	file_api_v1_chat_service_proto_depIdxs = nil
}